--- a/server/handlers/cacheHandlers.js
+++ b/server/handlers/cacheHandlers.js
@@ -3,16 +3,16 @@ const mcache = require('memory-cache');
 exports.responseCache = duration => (req, res, next ) => {
   const key = `__express__${req.originalURL || req.url}`;
   const cachedBody = mcache.get(key);
-  if (cachedBody)
-    res.send(cachedBody); 
+  if (cachedBody) res.send(cachedBody);
   else {
-    // destructuring statement below
-    // first element stores current .send method
-    // second element overrides current .send method with an arrow function
-    [res.sendResponse, res.send] = [res.send, (body) => {
-      mcache.put(key, body, duration); // store response in mcache
-      res.sendResponse(body); // call the orgininal .send method
-    }];
-    next();
+    const refreshResponse = (body) => {
+      mcache.put(key, body, duration); // cache body response
+      res.sendResponse(body); // call orgininal .send
+      // carefull calling property above pre defined
+    };
+
+    // store current .send method & then override .send method
+    [res.sendResponse, res.send] = [res.send, refreshResponse];
+    next(); // pass on to get uncached response
   }
-}
+};