--- a/server/server.js
+++ b/server/server.js
@@ -67,8 +67,8 @@ app.use((req, res, next) => {
 app.get(
   '/',
   responseCache(20 * 1000), // cache body data for 20 seconds
-  // async (req, res) => res.render('main', { countDown: 300 - await userController.cachedCountDown.getCount() }),
-     async (req, res) => res.render('main', { countDown: 300 - await userController.uncachedCountdown()}),
+  // previously below line got data from await userController.cachedCountDown.getCount()
+  async (req, res) => res.render('main', { countDown: 300 - await userController.uncachedCountdown()}),
 );
 
 // caching the reg get response