observational data

- zeit deployments looks like a decreased response times by 100 to 200 ms
- on local deployments more like a 500 to 600 ms decrease in response times
- and when under load will greatly reduce database calls

Testing still required by @Irev-Dev 
- do flashes still work where they are needed ???
- does the GET /reg url still work???