@mrhut10 2018-04-21
OK I should be able to fix that fairly easily if you haven't already, I'll close the request now and resubmit a fix. 
:)
Kind regards